requirements.txt
langchain
langchain-community
langchain-ollama
sentence-transformers
faiss-cpu
unstructured
python-docx
pypdf
langchain-huggingface
langchain-text-splitters
torch
accelerate